Dashboard Enhancment Structure: Projects, Tasks Pages - move rest of files.
