+++ src/app/(dashboard)/projects/[slug]/_components/ProjectProgress.jsx
