+++ src/app/(dashboard)/tasks/[slug]/details/page.jsx
@@ -1,15 +1,15 @@
 "use client"
 import Page from "@/components/Page.jsx";
 import SelectWithoutLabel from "@/components/Form/SelectWithoutLabel.jsx";
-import TasksList from "@/app/dashboard/projects/details/[slug]/components/TasksList.jsx";
-import TaskComments from "@/app/dashboard/projects/details/[slug]/components/TaskComments.jsx";
+import TasksList from "@/app/(dashboard)/projects/[slug]/_components/TasksList.jsx";
+import TaskComments from "@/app/(dashboard)/projects/[slug]/_components/TaskComments.jsx";
 import CommentInput from "@/components/CommentInput.jsx";
-import ProjectMembers from "@/app/dashboard/projects/details/[slug]/components/ProjectMembers.jsx";
-import AttachmentsList from "@/app/dashboard/projects/details/[slug]/components/AttachmentsList.jsx";
+import ProjectMembers from "@/app/(dashboard)/projects/[slug]/_components/ProjectMembers.jsx";
+import AttachmentsList from "@/app/(dashboard)/projects/[slug]/_components/AttachmentsList.jsx";
 import ActivityLogs from "@/components/ActivityLogs.jsx";
 import TimeLine from "@/components/TimeLine/TimeLine.jsx";
 import {useTranslation} from "react-i18next";
-import InfoCard from "@/app/dashboard/_components/InfoCard.jsx";
+import InfoCard from "@/app/(dashboard)/_components/InfoCard.jsx";
 import {useState} from "react";
 import {filterAndSortTasks} from "@/functions/functionsForTasks.js";
 import EditTaskModal from "@/app/(dashboard)/tasks/_modal/EditTaskModal.jsx";
