+++ src/app/(dashboard)/tasks/page.jsx
@@ -15,7 +15,7 @@ const NameAndDescription = dynamic(() => import("@/app/(dashboard)/projects/_com
 const AccountDetails = dynamic(() => import("@/app/(dashboard)/projects/_components/TableInfo/AccountDetails"), { ssr: false });
 const Priority = dynamic(() => import("@/app/(dashboard)/projects/_components/TableInfo/Priority"), { ssr: false });
 const Status = dynamic(() => import("@/app/(dashboard)/projects/_components/TableInfo/Status"), { ssr: false });
-const MembersListXLine = dynamic(() => import("@/app/dashboard/projects/details/[slug]/components/MembersListXLine"), { ssr: false });
+const MembersListXLine = dynamic(() => import("@/app/(dashboard)/projects/[slug]/_components/MembersListXLine"), { ssr: false });
 const TimeLine = dynamic(() => import("@/components/TimeLine/TimeLine"), { ssr: false });
 const EditTaskModal = dynamic(() => import("@/app/(dashboard)/tasks/_modal/EditTaskModal"), { ssr: false });
 const Alert = dynamic(() => import("@/components/Alert"), { ssr: false });
